requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
supabase>=2.0.0
pandas>=2.0.0
python-dateutil>=2.8.0
tqdm>=4.66.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0